Move PagesService into services package
